# docs/contracts/gateway.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WashGo Gateway/BFF foundation",
    "version": "1.0.0",
    "description": "Routing discovery only; does not assert business endpoints are implemented."
  },
  "paths": {
...
    "/api/v1/auth/session": {
      "get": {
        "operationId": "auth.session",
        "x-owning-service": "identity",
        "x-auth-enforcement-owner": "identity",
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "description": "A UUID is preserved; absent or invalid values are replaced by the gateway.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "header",
            "name": "traceparent",
            "required": false,
            "description": "Valid W3C version-00 trace context preserves the trace ID with a new span ID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owning service response; route-specific schema belongs to the owner.",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request UUID. External values are not trusted.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "X-Correlation-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "traceparent": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/GatewayError"
          },
          "401": {
            "$ref": "#/components/responses/GatewayError"
          },
          "403": {
            "$ref": "#/components/responses/GatewayError"
          },
          "404": {
            "$ref": "#/components/responses/GatewayError"
          },
          "409": {
            "$ref": "#/components/responses/GatewayError"
          },
          "413": {
            "$ref": "#/components/responses/GatewayError"
          },
          "422": {
            "$ref": "#/components/responses/GatewayError"
          },
          "429": {
            "$ref": "#/components/responses/GatewayError"
          },
          "500": {
            "$ref": "#/components/responses/GatewayError"
          },
          "502": {
            "$ref": "#/components/responses/GatewayError"
          },
          "503": {
            "$ref": "#/components/responses/GatewayError"
          },
          "504": {
            "$ref": "#/components/responses/GatewayError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-wg_access"
      }
    },
    "schemas": {
      "GatewayErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "requestId",
              "correlationId"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "REQUEST_INVALID",
                  "AUTH_REQUIRED",
                  "AUTH_FORBIDDEN",
                  "AUTH_CSRF",
                  "NOT_FOUND",
                  "CONFLICT",
                  "VALIDATION_FAILED",
                  "RATE_LIMITED",
                  "UPSTREAM_UNAVAILABLE",
                  "UPSTREAM_TIMEOUT",
                  "UPSTREAM_INVALID",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string",
                "description": "Fixed public text; never raw owner exceptions."
              },
              "requestId": {
                "type": "string",
                "format": "uuid"
              },
              "correlationId": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        }
      }
    },
    "responses": {
      "GatewayError": {
        "description": "Sanitized gateway error. No stack, SQL, cookies or credentials are returned.",
        "headers": {
          "X-Request-Id": {
            "description": "Server-generated request UUID. External values are not trusted.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          "X-Correlation-Id": {
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          "traceparent": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GatewayErrorEnvelope"
            }
          }
        }
      }
    }
  }
}